requirements.txt
python-dotenv
numpy<2
openai
openai-whisper